chore(demo): change examples order
