+++ projects/demo/src/pages/kit/date-time/date-time-mask-doc.template.html
@@ -48,6 +48,29 @@
         </tui-doc-example>
 
         <tui-doc-example
+            heading="Meridiem"
+            [content]="amPmExample"
+            [description]="amPmDescription"
+        >
+            <ng-template #amPmDescription>
+                Pass
+                <code>dayPeriod</code>
+                as a tuple of two locale-specific markers (e.g.
+                <code>['AM', 'PM']</code>
+                ,
+                <code>['上午', '下午']</code>
+                ) to switch the mask to a 12-hour time format with a meridiem suffix.
+
+                <p>
+                    Pass
+                    <code>['', '']</code>
+                    (default) to keep the 24-hour format.
+                </p>
+            </ng-template>
+            <date-time-mask-doc-example-2 />
+        </tui-doc-example>
+
+        <tui-doc-example
             heading="Locale"
             [content]="localeExample"
             [description]="localeDescription"
@@ -67,7 +90,7 @@
                 one.
             </ng-template>
 
-            <date-time-mask-doc-example-6 />
+            <date-time-mask-doc-example-3 />
         </tui-doc-example>
 
         <tui-doc-example
@@ -80,7 +103,7 @@
                 <code>dateTimeSeparator</code>
                 parameter to configure separator between date and time strings.
             </ng-template>
-            <date-time-mask-doc-example-2 />
+            <date-time-mask-doc-example-4 />
         </tui-doc-example>
 
         <tui-doc-example
@@ -104,7 +127,7 @@
                 </a>
                 .
             </ng-template>
-            <date-time-mask-doc-example-3 />
+            <date-time-mask-doc-example-5 />
         </tui-doc-example>
 
         <tui-doc-example
@@ -129,30 +152,7 @@
                     (default value) to disable this feature.
                 </p>
             </ng-template>
-            <date-time-mask-doc-example-4 />
-        </tui-doc-example>
-
-        <tui-doc-example
-            heading="Meridiem"
-            [content]="amPmExample"
-            [description]="amPmDescription"
-        >
-            <ng-template #amPmDescription>
-                Pass
-                <code>dayPeriod</code>
-                as a tuple of two locale-specific markers (e.g.
-                <code>['AM', 'PM']</code>
-                ,
-                <code>['上午', '下午']</code>
-                ) to switch the mask to a 12-hour time format with a meridiem suffix.
-
-                <p>
-                    Pass
-                    <code>['', '']</code>
-                    (default) to keep the 24-hour format.
-                </p>
-            </ng-template>
-            <date-time-mask-doc-example-5 />
+            <date-time-mask-doc-example-6 />
         </tui-doc-example>
     </ng-template>
