+++ projects/demo/src/pages/kit/date-time/examples/3-locale/mask.ts
