+++ projects/demo/src/pages/kit/date-time/examples/4-date-time-separator/mask.ts
