+++ projects/demo/src/pages/kit/date-time/examples/5-min-max/mask.ts
