+++ projects/demo/src/pages/kit/date-time/examples/6-time-step/component.ts
@@ -6,7 +6,7 @@ import {TuiInput} from '@taiga-ui/core';
 import mask from './mask';
 
 @Component({
-    selector: 'date-time-mask-doc-example-3',
+    selector: 'date-time-mask-doc-example-6',
     imports: [FormsModule, MaskitoDirective, TuiInput],
     template: `
         <tui-textfield
@@ -14,7 +14,7 @@ import mask from './mask';
             [style.max-width.rem]="30"
             [filler]="filler"
         >
-            <label tuiLabel>Min-max</label>
+            <label tuiLabel>Time Stepping</label>
             <input
                 inputmode="decimal"
                 tuiInput
@@ -25,8 +25,8 @@ import mask from './mask';
     `,
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class DateTimeMaskDocExample3 {
-    protected value = '09-01-2018, 15:30';
-    protected readonly filler = 'dd-mm-yyyy, hh:mm';
+export class DateTimeMaskDocExample6 {
+    protected value = '09.01.2018, 15:30';
+    protected readonly filler = 'dd.mm.yyyy, hh:mm';
     protected readonly mask = mask;
 }
