+++ projects/demo/src/pages/kit/date-time/examples/6-time-step/mask.ts
